fix: commit index only when the event index is updated
The previous logic caused cache inserts even when the last event index
was not updated (ie. no new events in the stream), causing unnecessary
cache calls. the new logic maintains an in-memory map with the latest
event index and only updates the cache if a new event index is found.

Signed-off-by: Chinmay D. Pai <chinmaydpai@gmail.com>